backend/requirements.txt
# Cinematic Music Platform Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# HTTP Client
httpx>=0.26.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database (optional, for production)
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Redis (for task queue and caching)
redis>=5.0.0
celery>=5.3.0

# WebSocket support
websockets>=12.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
numpy>=1.26.0
scipy>=1.11.0

# Machine Learning
torch>=2.0.0
transformers>=4.35.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
aiohttp>=3.9.0

# Rate limiting
slowapi>=0.1.8

# Logging
loguru>=0.7.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0

